# [block] add support for notification suppression