--- a/src/vmm/src/persist.rs
+++ b/src/vmm/src/persist.rs
@@ -19,6 +19,8 @@ use crate::vstate::{self, vcpu::VcpuState, vm::VmState};
 use crate::device_manager::persist::DeviceStates;
 use crate::memory_snapshot;
 use crate::memory_snapshot::{GuestMemoryState, SnapshotMemory};
+#[cfg(target_arch = "x86_64")]
+use crate::version_map::FC_V0_23_SNAP_VERSION;
 use crate::version_map::FC_VERSION_TO_SNAP_VERSION;
 use crate::{Error as VmmError, EventManager, Vmm};
 #[cfg(target_arch = "x86_64")]
@@ -34,8 +36,6 @@ use versionize::{VersionMap, Versionize, VersionizeResult};
 use versionize_derive::Versionize;
 use vm_memory::GuestMemoryMmap;
 
-#[cfg(target_arch = "x86_64")]
-const FC_V0_23_SNAP_VERSION: u16 = 1;
 #[cfg(target_arch = "x86_64")]
 const FC_V0_23_MAX_DEVICES: u32 = 11;
 
@@ -297,23 +297,25 @@ fn snapshot_memory_to_file(
 
 /// Validate the microVM version and translate it to its corresponding snapshot data format.
 pub fn get_snapshot_data_version(
-    version: &Option<String>,
+    maybe_fc_version: &Option<String>,
     version_map: &VersionMap,
     _vmm: &Vmm,
 ) -> std::result::Result<u16, CreateSnapshotError> {
-    if version.is_none() {
-        return Ok(version_map.latest_version());
+    let fc_version = match maybe_fc_version {
+        None => return Ok(version_map.latest_version()),
+        Some(version) => version,
+    };
+    validate_fc_version_format(fc_version)?;
+    let data_version = *FC_VERSION_TO_SNAP_VERSION
+        .get(fc_version)
+        .ok_or(CreateSnapshotError::UnsupportedVersion)?;
+
+    #[cfg(target_arch = "x86_64")]
+    if data_version <= FC_V0_23_SNAP_VERSION {
+        validate_devices_number(_vmm.mmio_device_manager.used_irqs_count())?;
     }
-    validate_fc_version_format(version.as_ref().unwrap())?;
-    match FC_VERSION_TO_SNAP_VERSION.get(version.as_ref().unwrap()) {
-        #[cfg(target_arch = "x86_64")]
-        Some(&FC_V0_23_SNAP_VERSION) => {
-            validate_devices_number(_vmm.mmio_device_manager.used_irqs_count())?;
-            Ok(FC_V0_23_SNAP_VERSION)
-        }
-        Some(data_version) => Ok(*data_version),
-        _ => Err(CreateSnapshotError::UnsupportedVersion),
-    }
+
+    Ok(data_version)
 }
 
 /// Validates that snapshot CPU vendor matches the host CPU vendor.