… Cargo.lock
… src/jailer/Cargo.toml
… src/vmm/Cargo.toml
… src/vmm/src/device_manager/mmio.rs
… src/vmm/src/version_map.rs
… tests/framework/artifacts.py
… tests/framework/utils.py
… tests/integration_tests/build/test_binary_size.py
… tests/integration_tests/build/test_coverage.py
… tests/integration_tests/functional/test_cmd_line_parameters.py
… tests/integration_tests/functional/test_snapshot_advanced.py
… tests/integration_tests/functional/test_snapshot_version.py
… tests/integration_tests/performance/test_snapshot_perf.py