# Reason for This PR

#2635 

## Description of Changes

### Performance Results

**Test setup:** We run on a m5d.metal host with Ubuntu and kernel 5.8.18. We start a number of microVMs with 8vCPUs and 16GB of RAM running linux kernel v5.8.18 as a guest OS. Each microVMs is pinned to NUMA node 1. We mount a ramdisk and we pin it to NUMA node 1 as well. For each microVM we create an ext4 file on the ramdisk and we attach it to the microVM as /dev/vdb. We want to have the content of the drives in the page cache in order to asses the performance improvement brought by notification suppression. Otherwise we would be limited by the disk access speed. Inside the VM we run the following command:

```
fio --name=test --ioengine=libaio --rw=readwrite --randrepeat=1 --bs=4k --size=500m --numjobs=1 --iodepth=32 --runtime=30 --time_base=1 --end_fsync=1 --direct=1 --filename=/dev/vdb
```

Current implementation:
|description|IOPS|VMM thread CPU load|kworker threads CPU load|IOPS per CPU load|
|-|-|-|-|-|
|1 VMs|104.86|0.36|0.13|213.5|
|4 VMs|439.28|1.51|0.52|216.34|
|8 VMs|849.83|3.1|1.07|203.47|
|16 VMs|1524.45|6.76|2.24|169.51|
|32 VMs|2202.07|8.32|2.19|209.51|
|64 VMs|1182.12|5.76|1.77|156.89|

Notification suppression:
|description|IOPS|VMM thread CPU load|kworker threads CPU load|IOPS per CPU load|increase in efficiency|
|-|-|-|-|-|-|
|1 VMs|160.72|0.42|0.09|315.65|47.85%|
|4 VMs|599.36|1.64|0.36|299.79|38.57%|
|8 VMs|1175.92|3.33|0.81|283.79|39.48%|
|16 VMs|1872.87|7.07|2.12|203.68|20.16%|
|32 VMs|2551.24|8.67|2.61|226.2|7.97%|
|64 VMs|1369.84|6.07|1.79|174.26|11.07%|


- [ ] This functionality can be added in [`rust-vmm`](https://github.com/rust-vmm).

## License Acceptance

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.

## PR Checklist

`[Author TODO: Meet these criteria.]`
`[Reviewer TODO: Verify that these criteria are met. Request changes if not]`

- [x] All commits in this PR are signed (`git commit -s`).
- [x] The reason for this PR is clearly provided (issue no. or explanation).
- [x] The description of changes is clear and encompassing.
- [x] Any required documentation changes (code and docs) are included in this PR.
- [x] Any newly added `unsafe` code is properly documented.
- [x] Any API changes are reflected in `firecracker/swagger.yaml`.
- [x] Any user-facing changes are mentioned in `CHANGELOG.md`.
- [x] All added/changed functionality is tested.