@serban300 2021-08-05
I made some small adjustments to the notification suppression functionality and also implemented the logic that breaks snapshot compatibility with previous versions of Firecracker. The PR is now ready for review. @acatangiu @sandreim @firecracker-microvm/compute-capsule PTAL when you have time !